Update CourseFormater to output classes as specified in skel-turmas.json
The only exception to the skel-turmas.json format is that the course attributes 'numeroDeSolicitacoes' and 'numeroDeVagas' aren't passed along to the classes objects. Because they are needed only to calculate the 'concorrencia' attribute.

Only the function 'parseClasses(..)' needs to be exposed, so that's why the other functions aren't a member of the CourseFormate object.