fix(ResultsPage): load more and error notification
This commit fixes the load more functionality by spliting it from the first search, that is handled in the effect hook. This way, the results are updated correctly when a different search is submited by the user, given that  the user has clicked in the load more button.

This commit also renders the results unconditionally and shows the load more animation inside the button. Now the results that have been already loaded won't disappear when fetching more results.

Also, this commit sets a max width and height in the loading animation so that it will fit correctly inside a button.